Sync between Particle and ParticleCloud
Now, both of Particle and ParticleCloud share their attributes. It is still a work in progress, as in their current version they both have to support all entries in the Attr enum. Otherwise things might get messy.

Note that it is way over engineered for what it is meant to be doing!